refactor(heki): the store moves outside its folder and splits codec from journal
adapters/driven/heki.rb is now the repository surface; the binary
snapshot codec (magic-header framing over deflated, id-sorted JSON)
lives in heki/snapshot.rb and the fsynced append-only journal in
heki/journal.rb. The .adapter declaration moves up beside the class,
matching memory/folder/prism from the earlier phase.
