+++ lib/hecksagain/adapters/driven.rb
@@ -11,6 +11,6 @@ end
 require_relative "driven/memory"
 require_relative "driven/sqlite/sqlite"
 require_relative "driven/postgres/postgres"
-require_relative "driven/heki/heki"
+require_relative "driven/heki"
 require_relative "driven/prism"
 require_relative "driven/folder"
