+++ lib/hecksagain/adapters/driven/heki.rb
@@ -1,10 +1,18 @@
 require "json"
-require "zlib"
 require "fileutils"
+require_relative "heki/snapshot"
+require_relative "heki/journal"
 
 module Hecksagain
   module Adapters
+    # The file-backed store: a compressed snapshot (heki/snapshot.rb)
+    # plus an append-only journal beside it (heki/journal.rb). What stays
+    # here is the repository surface — find/all/save/delete, and the
+    # entry append/project pair the persistence port drives.
     class Heki
+      include Snapshot
+      include Journal
+
       MAGIC = "HEKI"
       HEADER_BYTES = 8
 
@@ -54,16 +62,6 @@ module Hecksagain
         entry
       end
 
-      def entries
-        return [] unless File.exist?(@journal_path)
-
-        File.readlines(@journal_path, chomp: true).reject(&:empty?).map do |line|
-          value = JSON.parse(line)
-          state = value["state"]&.transform_keys(&:to_sym)
-          Ports::Persistence::Entry.new(operation: value.fetch("operation"), id: value.fetch("id"), state: state, mirrors: value["mirrors"])
-        end
-      end
-
       def save(instance)
         entry = Ports::Persistence::Entry.new(operation: "save", id: instance.id.to_s, state: instance.state.dup)
         append(entry)
@@ -103,62 +101,11 @@ module Hecksagain
         replay_journal(snapshot)
       end
 
-      def read_snapshot
-        return {} unless File.exist?(@path)
-
-        data = File.binread(@path)
-        raise Malformed, "#{@path}: too short" if data.bytesize < HEADER_BYTES
-        raise Malformed, "#{@path}: bad magic" unless data[0, 4] == MAGIC
-
-        JSON.parse(Zlib::Inflate.inflate(data[HEADER_BYTES..]))
-      rescue Zlib::DataError => e
-        raise Malformed, "#{@path}: zlib error: #{e.message}"
-      rescue JSON::ParserError => e
-        raise Malformed, "#{@path}: json error: #{e.message}"
-      end
-
-      def replay_journal(records)
-        return records unless File.exist?(@journal_path)
-
-        File.foreach(@journal_path) do |line|
-          entry = JSON.parse(line)
-          id    = entry.fetch("id")
-
-          case entry.fetch("operation")
-          when "save"   then records[id] = entry.fetch("state")
-          when "delete" then records.delete(id)
-          else raise Malformed, "#{@journal_path}: unknown journal operation #{entry.fetch("operation").inspect}"
-          end
-        end
-        records
-      rescue JSON::ParserError => e
-        raise Malformed, "#{@journal_path}: json error: #{e.message}"
-      end
-
-      def write(records)
-        sorted = records.sort_by { |id, _| id }.to_h
-        json   = JSON.generate(sorted)
-
-        File.binwrite(
-          @path,
-          MAGIC + [sorted.size].pack("N") + Zlib::Deflate.deflate(json, Zlib::BEST_COMPRESSION)
-        )
-      end
-
       def resolve_path(settings, root)
         declared = settings[:dir] || settings["dir"] || "data"
         dir      = declared.start_with?("/") ? declared : File.join(root || Dir.pwd, declared)
 
         File.join(dir, "#{@aggregate.storage_name}.heki")
-      end
-
-      def append_entry(operation, id, state)
-        File.open(@journal_path, "ab") do |journal|
-          journal.write(JSON.generate(operation: operation, id: id.to_s, state: state, mirrors: @entry_mirrors))
-          journal.write("\n")
-          journal.flush
-          journal.fsync
-        end
       end
     end
   end
