+++ lib/hecksagain/adapters/driven/heki/journal.rb
@@ -1,0 +1,48 @@
+module Hecksagain
+  module Adapters
+    class Heki
+      # The append-only journal beside the snapshot: one JSON line per
+      # entry, fsynced on append, replayed over the snapshot on read.
+      module Journal
+        def entries
+          return [] unless File.exist?(@journal_path)
+
+          File.readlines(@journal_path, chomp: true).reject(&:empty?).map do |line|
+            value = JSON.parse(line)
+            state = value["state"]&.transform_keys(&:to_sym)
+            Ports::Persistence::Entry.new(operation: value.fetch("operation"), id: value.fetch("id"), state: state, mirrors: value["mirrors"])
+          end
+        end
+
+        private
+
+        def replay_journal(records)
+          return records unless File.exist?(@journal_path)
+
+          File.foreach(@journal_path) do |line|
+            entry = JSON.parse(line)
+            id    = entry.fetch("id")
+
+            case entry.fetch("operation")
+            when "save"   then records[id] = entry.fetch("state")
+            when "delete" then records.delete(id)
+            else raise Malformed, "#{@journal_path}: unknown journal operation #{entry.fetch("operation").inspect}"
+            end
+          end
+          records
+        rescue JSON::ParserError => e
+          raise Malformed, "#{@journal_path}: json error: #{e.message}"
+        end
+
+        def append_entry(operation, id, state)
+          File.open(@journal_path, "ab") do |journal|
+            journal.write(JSON.generate(operation: operation, id: id.to_s, state: state, mirrors: @entry_mirrors))
+            journal.write("\n")
+            journal.flush
+            journal.fsync
+          end
+        end
+      end
+    end
+  end
+end
